float to double and raycast loop breaker

+++ c3d_math/c3d_math.h
@@ -6,7 +6,7 @@
 /*   By: zanikin <zanikin@student.42yerevan.am>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/12 06:41:15 by zanikin           #+#    #+#             */
-/*   Updated: 2024/10/03 16:06:11 by zanikin          ###   ########.fr       */
+/*   Updated: 2024/10/09 20:00:09 by zanikin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,12 +21,12 @@
 
 void	free_mat(t_mat *mat);
 bool	create_mat(t_mat *mat, unsigned int w, unsigned int h, t_err *error);
-float	square_distance(const t_vec *p1, const t_vec *p2);
-float	distance(const t_vec *p1, const t_vec *p2);
-void	set_vec(t_vec *v, float x, float y);
+double	square_distance(const t_vec *p1, const t_vec *p2);
+double	distance(const t_vec *p1, const t_vec *p2);
+void	set_vec(t_vec *v, double x, double y);
 void	set_ivec(t_ivec *v, int x, int y);
 void	vec_add(const t_vec *v1, const t_vec *v2, t_vec *rv);
 void	vec_sub(const t_vec *v1, const t_vec *v2, t_vec *rv);
-void	vec_mul(const t_vec *v, float a, t_vec *rv);
-void	vec_div(const t_vec *v, float a, t_vec *rv);
+void	vec_mul(const t_vec *v, double a, t_vec *rv);
+void	vec_div(const t_vec *v, double a, t_vec *rv);
 #endif
