+++ game/hooks.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   hooks.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mamazari <mamazari@student.42.fr>          +#+  +:+       +#+        */
+/*   By: zanikin <zanikin@student.42yerevan.am>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/04 15:23:15 by zanikin           #+#    #+#             */
-/*   Updated: 2024/10/04 16:29:01 by mamazari         ###   ########.fr       */
+/*   Updated: 2024/10/09 20:00:14 by zanikin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,8 +24,8 @@ void		render(t_game *game);
 void		free_game(t_render *r, t_mat *map, t_mat *states);
 int			exit_game(t_game *game);
 
-static void	rotate(t_game *game, float cos_v, float sin_v);
-static void	move(t_game *game, float mdx, float mdy);
+static void	rotate(t_game *game, double cos_v, double sin_v);
+static void	move(t_game *game, double mdx, double mdy);
 
 int	mouse_look(t_game *game)
 {
@@ -61,9 +61,9 @@ int	key_hook(int keycode, t_game *game)
 	return (0);
 }
 
-static void	rotate(t_game *game, float cos_v, float sin_v)
+static void	rotate(t_game *game, double cos_v, double sin_v)
 {
-	float	x;
+	double	x;
 
 	x = game->prot.x;
 	game->prot.x = cos_v * x - sin_v * game->prot.y;
@@ -74,7 +74,7 @@ static void	rotate(t_game *game, float cos_v, float sin_v)
 	render(game);
 }
 
-static void	move(t_game *game, float mdx, float mdy)
+static void	move(t_game *game, double mdx, double mdy)
 {
 	game->ppos.x += mdx * MOVEMENT_RESOLUTION;
 	game->ppos.y += mdy * MOVEMENT_RESOLUTION;
